+++ test/message-size.js
@@ -12,7 +12,7 @@ function createMessage(transferCount) {
       sendBuffer: new Buffer([0x01, 0xc0, 0x00]),
       receiveBuffer: new Buffer(3),
       byteLength: 3,
-      speed: 1350000,
+      speedHz: 1350000,
       chipSelectChange: true
     });
   }