+++ test/stress/tmp36-sync.js
@@ -11,7 +11,7 @@ var message = [{
   sendBuffer: new Buffer([0x01, 0xd0, 0x00]),
   receiveBuffer: new Buffer([0, 0, 0]),
   byteLength: 3,
-  speed: 20000
+  speedHz: 20000
 }];
 
 while (true) {