+++ test/stress/tmp36.js
@@ -13,7 +13,7 @@ mcp3008 = spi.open(0, 0, {mode: spi.MODE0}, function (err) {
     sendBuffer: new Buffer([0x01, 0xd0, 0x00]),
     receiveBuffer: new Buffer([0, 0, 0]),
     byteLength: 3,
-    speed: 20000
+    speedHz: 20000
   }];
 
   assert(!err, 'can\'t open device');